Abstract out text normalization into own file
