+++ client/src/components/BlockEditor/BlockEditorEditing.tsx
@@ -13,19 +13,7 @@ import { valueToDatabaseJSON } from "../../lib/slateParser";
 import { exportSelection } from "../../modules/blockEditor/actions";
 import * as _ from "lodash";
 import { UPDATE_BLOCKS } from "../../graphqlQueries";
-import { SPACER } from "../../lib/slate-pointers/exportedPointerSpacer";
-
-interface Change {
-  insertTextByKey: any;
-  value: any;
-}
-
-interface TextNode {
-  key: string;
-  object: string;
-  leaves: any[];
-  text: string;
-}
+import { normalizeChangeWrtTextNodeSpacing } from "./normalizeChange";
 
 const BlockEditorStyle = styled.div`
   background: #f4f4f4;
@@ -136,60 +124,8 @@ export class BlockEditorEditingPresentational extends React.Component<
     }
   }
 
-  public normalizeChangeWrtTextNodeSpacing(c: Change) {
-    const value = c.value;
-
-    function directlyBeforeExport(textNode: TextNode) {
-      return value.document.getNextSibling(textNode.key) && value.document.getNextSibling(textNode.key).type === "pointerExport";
-    }
-
-    function directlyAfterExport(textNode: TextNode) {
-      return value.document.getPreviousSibling(textNode.key) && value.document.getPreviousSibling(textNode.key).type === "pointerExport";
-    }
-
-    function insideExport(textNode: TextNode) {
-      return value.document.getParent(textNode.key) && value.document.getParent(textNode.key).type === "pointerExport";
-    }
-
-    function firstCharIsSpacer(textNode: TextNode) {
-      return textNode.text.charAt(0) === SPACER;
-    }
-
-    function lastCharIsSpacer(textNode: TextNode) {
-      return textNode.text.charAt(textNode.text.length - 1) === SPACER;
-    }
-
-    value.document.getTexts().forEach(textNode => {
-      if (!insideExport(textNode) && directlyBeforeExport(textNode) && !lastCharIsSpacer(textNode)) {
-        c.insertTextByKey(textNode.key, textNode.text.length, SPACER);
-      }
-
-      if (!insideExport(textNode) && directlyAfterExport(textNode) && !firstCharIsSpacer(textNode)) {
-        c.insertTextByKey(textNode.key, 0, SPACER);
-      }
-
-      if (insideExport(textNode) && !firstCharIsSpacer(textNode)) {
-        c.insertTextByKey(textNode.key, 0, SPACER);
-      }
-
-      if (insideExport(textNode) && !lastCharIsSpacer(textNode)) {
-        c.insertTextByKey(textNode.key, textNode.text.length, SPACER);
-      }
-
-      // for edge case where there is only two spaces in a pointer
-      // and you delete one, then first and last char are both spaces
-      // since they are the same char, so you need to additionaly enforce
-      // that there are at least two chars
-      if (insideExport(textNode) && textNode.text.length === 1) {
-        c.insertTextByKey(textNode.key, 0, SPACER);
-      }
-    });
-
-    return c;
-  }
-
   public componentDidMount() {
-    const change = this.normalizeChangeWrtTextNodeSpacing(this.props.block.value.change());
+    const change = normalizeChangeWrtTextNodeSpacing(this.props.block.value.change());
     this.props.updateBlock({ id: this.props.block.id, value: change.value, pointerChanged: false });
   }
 
@@ -334,7 +270,7 @@ export class BlockEditorEditingPresentational extends React.Component<
     // Slate uses immutable objects, if it has changed then normalize wrt
     // pointer spacing
     if (c.value.document !== this.props.block.value.document) {
-      c = this.normalizeChangeWrtTextNodeSpacing(c);
+      c = normalizeChangeWrtTextNodeSpacing(c);
     }
 
     // the rest of this method before the last two lines are trying to prevent
